Fixes Gregor's suggestions
